Refactor models and imports, split socket events into separate files
