+++ frontend/src/app/models/index.ts
@@ -1,15 +1,20 @@
 // Question and Quiz related interfaces
 export * from './question.model';
+export * from './question-submission.model';
+export * from './question-result.model';
 export * from './quiz-session.model';
 export * from './quiz-state.model';
 export * from './sample-question.model';
 
 // Team and User interfaces
 export * from './team.model';
+export * from './leaderboard-team.model';
 export * from './user-session.model';
 
 // Answer related interfaces
 export * from './answer.model';
+export * from './sequence-answer.model';
+export * from './review-answer.model';
 
 // Request interfaces
 export * from './create-question-request.model';
@@ -23,4 +28,6 @@ export * from './load-sample-questions-request.model';
 export * from './join-session-data.model';
 export * from './submit-answer-data.model';
 export * from './presenter-action.model';
-export * from './socket-events.model';
+export * from './team-events.model';
+export * from './question-events.model';
+export * from './session-events.model';
