+++ frontend/src/app/models/leaderboard-team.model.ts
@@ -1,0 +1,7 @@
+export interface LeaderboardTeam {
+  id: string;
+  name: string;
+  total_points: number;
+  answers_submitted: number;
+  correct_answers: number;
+}
