+++ frontend/src/app/models/question-result.model.ts
@@ -1,0 +1,9 @@
+export interface QuestionResult {
+  question_id: number;
+  team_id: number;
+  answer: string | string[];
+  correct_answer: string | string[];
+  is_correct: boolean;
+  points_earned: number;
+  time_taken: number;
+}
