+++ frontend/src/app/models/question-submission.model.ts
@@ -1,0 +1,7 @@
+export interface QuestionSubmission {
+  question_id: number;
+  team_id: number;
+  answer: string | string[];
+  submitted_at: Date;
+  time_taken?: number;
+}
