+++ frontend/src/app/models/question.model.ts
@@ -17,21 +17,3 @@ export interface Question {
   created_at: string;
   updated_at: string;
 }
-
-export interface QuestionSubmission {
-  question_id: number;
-  team_id: number;
-  answer: string | string[];
-  submitted_at: Date;
-  time_taken?: number;
-}
-
-export interface QuestionResult {
-  question_id: number;
-  team_id: number;
-  answer: string | string[];
-  correct_answer: string | string[];
-  is_correct: boolean;
-  points_earned: number;
-  time_taken: number;
-}
