+++ frontend/src/app/models/sequence-answer.model.ts
@@ -1,0 +1,6 @@
+export interface SequenceAnswer {
+  id: string;
+  answer_id: string;
+  item_text: string;
+  position: number;
+}
