+++ frontend/src/app/models/session-events.model.ts
@@ -1,0 +1,15 @@
+export interface LeaderboardUpdatedEvent {
+  teams: any[];
+}
+
+export interface RoundStartedEvent {
+  roundNumber: number;
+}
+
+export interface SessionEndedEvent {
+  teams: any[];
+}
+
+export interface SessionEndedErrorEvent {
+  message: string;
+}
