+++ frontend/src/app/models/team-events.model.ts
@@ -1,0 +1,17 @@
+export interface TeamJoinedEvent {
+  teamId: string;
+  teamName: string;
+  sessionStatus: string;
+}
+
+export interface TeamJoinedSessionEvent {
+  teamId: string;
+  teamName: string;
+}
+
+export interface ExistingTeamsEvent {
+  teams: Array<{
+    id: string;
+    name: string;
+  }>;
+}
